+++ tensorflow/contrib/tensorrt/BUILD
@@ -344,22 +344,22 @@ py_test(
 py_test(
     name = "converter_unit_tests",
     srcs = [
-        "test/unit_tests/base_unit_test.py",
-        "test/unit_tests/BatchMatMulTest.py",
-        "test/unit_tests/BiasaddMatMulTest.py",
-        "test/unit_tests/BinaryTensorWeightBroadcastTest.py",
-        "test/unit_tests/ConcatenationTest.py",
-        "test/unit_tests/ConstBroadcastTest.py",
-        "test/unit_tests/MultiConnectionNeighborEngineTest.py",
-        "test/unit_tests/NeighboringEngineTest.py",
-        "test/unit_tests/run_test.py",
-        "test/unit_tests/UnaryTest.py",
-        "test/unit_tests/unit_tests.py",
-        "test/unit_tests/utilities.py",
-        "test/unit_tests/VGGBlockNCHWTest.py",
-        "test/unit_tests/VGGBlockTest.py",
+        "test/base_unit_test.py",
+        "test/batch_matmul_test.py",
+        "test/biasadd_matmul_test.py",
+        "test/binary_tensor_weight_broadcast_test.py",
+        "test/concatenation_test.py",
+        "test/const_broadcast_test.py",
+        "test/multi_connection_neighbor_engine_test.py",
+        "test/neighboring_engine_test.py",
+        "test/run_test.py",
+        "test/unary_test.py",
+        "test/unit_tests.py",
+        "test/utilities.py",
+        "test/vgg_block_nchw_test.py",
+        "test/vgg_block_test.py",
     ],
-    main = "test/unit_tests/unit_tests.py",
+    main = "test/unit_tests.py",
     srcs_version = "PY2AND3",
     tags = [
         "manual",
