+++ tensorflow/contrib/tensorrt/test/multi_connection_neighbor_engine_test.py
@@ -18,30 +18,17 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import argparse
 import numpy as np
 
-from tensorflow.contrib import tensorrt as trt
-from tensorflow.core.protobuf import config_pb2 as cpb2
-from tensorflow.core.protobuf import rewriter_config_pb2 as rwpb2
-from tensorflow.python.client import session as csess
+from tensorflow.core.protobuf import config_pb2
 from tensorflow.python.framework import constant_op
 from tensorflow.python.framework import dtypes
-from tensorflow.python.framework import importer as importer
-from tensorflow.python.framework import ops as ops
+from tensorflow.python.framework import ops
 from tensorflow.python.ops import array_ops
 from tensorflow.python.ops import nn
-from tensorflow.python.ops import nn_ops
-from tensorflow.python.ops import nn_impl
-from tensorflow.python.ops import variable_scope
-from tensorflow.python.ops import variables
-from tensorflow.python.ops import init_ops
-from tensorflow.python.ops import gen_array_ops
 from tensorflow.python.ops import gen_math_ops
 from tensorflow.python.ops import math_ops
-from tensorflow.python.training import training
-from tensorflow.contrib.tensorrt.test.unit_tests.base_unit_test import BaseUnitTest
-from tensorflow.contrib.tensorrt.test.unit_tests.utilities import get_all_variables
+from tensorflow.contrib.tensorrt.test.base_unit_test import BaseUnitTest
 
 
 class MultiConnectionNeighborEngineTest(BaseUnitTest):
@@ -63,8 +50,8 @@ class MultiConnectionNeighborEngineTest(BaseUnitTest):
 
   def neighboring_tensor_test(self):
     g = ops.Graph()
-    gpu_options = cpb2.GPUOptions(per_process_gpu_memory_fraction=0.50)
-    sessconfig = cpb2.ConfigProto(gpu_options=gpu_options)
+    gpu_options = config_pb2.GPUOptions(per_process_gpu_memory_fraction=0.50)
+    sessconfig = config_pb2.ConfigProto(gpu_options=gpu_options)
     with g.as_default():
       x = array_ops.placeholder(
           dtype=dtypes.float32, shape=self.inp_dims, name="input")
