+++ tensorflow/contrib/tensorrt/test/vgg_block_nchw_test.py
@@ -18,75 +18,68 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import argparse
 import numpy as np
 
-from tensorflow.contrib import tensorrt as trt
-from tensorflow.core.protobuf import config_pb2 as cpb2
-from tensorflow.core.protobuf import rewriter_config_pb2 as rwpb2
-from tensorflow.python.client import session as csess
+from tensorflow.core.protobuf import config_pb2
 from tensorflow.python.framework import constant_op
 from tensorflow.python.framework import dtypes
-from tensorflow.python.framework import importer as importer
-from tensorflow.python.framework import ops as ops
+from tensorflow.python.framework import ops
 from tensorflow.python.ops import array_ops
 from tensorflow.python.ops import nn
 from tensorflow.python.ops import nn_ops
 from tensorflow.python.ops import nn_impl
-from tensorflow.python.ops import variable_scope
-from tensorflow.python.ops import variables
-from tensorflow.python.ops import init_ops
-from tensorflow.python.ops import gen_array_ops
-from tensorflow.python.ops import gen_math_ops
-from tensorflow.python.ops import math_ops
-from tensorflow.python.training import training
-from tensorflow.contrib.tensorrt.test.unit_tests.base_unit_test import BaseUnitTest
-from tensorflow.contrib.tensorrt.test.unit_tests.utilities import get_all_variables
+from tensorflow.contrib.tensorrt.test.base_unit_test import BaseUnitTest
 
 
-class NeighboringEngineTest(BaseUnitTest):
-  """Neighboring node wiring tests in TF-TRT conversion"""
+class VGGBlockNCHWTest(BaseUnitTest):
+  """single vgg layer in NCHW unit tests in TF-TRT"""
 
   def __init__(self, log_file='log.txt'):
-    super(NeighboringEngineTest, self).__init__()
+    super(VGGBlockNCHWTest, self).__init__()
     self.static_mode_list = {"FP32", "FP16"}
     self.debug = True
     self.dynamic_mode_list = {}
-    self.inp_dims = (2, 3, 7, 5)
+    self.inp_dims = (5, 2, 8, 8)
     self.dummy_input = np.random.random_sample(self.inp_dims)
-    self.get_network = self.neighboring_tensor_test
-    self.expect_nb_nodes = 5
+    self.get_network = self.get_simple_graph_def
+    self.expect_nb_nodes = 3
     self.log_file = log_file
     self.test_name = self.__class__.__name__
-    self.allclose_rtol = 0.05
-    self.allclose_atol = 0.05
 
-  def neighboring_tensor_test(self):
+  def get_simple_graph_def(self):
     g = ops.Graph()
-    gpu_options = cpb2.GPUOptions(per_process_gpu_memory_fraction=0.50)
-    sessconfig = cpb2.ConfigProto(gpu_options=gpu_options)
+    gpu_options = config_pb2.GPUOptions(per_process_gpu_memory_fraction=0.50)
+    sessconfig = config_pb2.ConfigProto(gpu_options=gpu_options)
     with g.as_default():
       x = array_ops.placeholder(
           dtype=dtypes.float32, shape=self.inp_dims, name="input")
+      x, mean_x, var_x = nn_impl.fused_batch_norm(
+          x,
+          np.random.randn(2).astype(np.float32),
+          np.random.randn(2).astype(np.float32),
+          mean=np.random.randn(2).astype(np.float32),
+          variance=np.random.randn(2).astype(np.float32),
+          data_format="NCHW",
+          is_training=False)
       e = constant_op.constant(
-          np.random.normal(.3, 0.05, [3, 2, 3, 4]),
-          name="weights",
-          dtype=dtypes.float32)
+          np.random.randn(1, 1, 2, 6), name="weights", dtype=dtypes.float32)
       conv = nn.conv2d(
           input=x,
           filter=e,
           data_format="NCHW",
-          strides=[1, 1, 1, 1],
-          padding="VALID",
+          strides=[1, 1, 2, 2],
+          padding="SAME",
           name="conv")
       b = constant_op.constant(
-          np.random.normal(1.0, 1.0, [1, 4, 1, 1]),
-          name="bias",
-          dtype=dtypes.float32)
-      t = conv * b
-
-      e = gen_math_ops.tan(conv)
-      t = t - e
-      array_ops.squeeze(t, name="output")
+          np.random.randn(6), name="bias", dtype=dtypes.float32)
+      t = nn.bias_add(conv, b, data_format="NCHW", name="biasAdd")
+      relu = nn.relu(t, "relu")
+      idty = array_ops.identity(relu, "ID")
+      v = nn_ops.max_pool(
+          idty, [1, 1, 2, 2], [1, 1, 2, 2],
+          "VALID",
+          data_format="NCHW",
+          name="max_pool")
+      array_ops.squeeze(v, name="output")
 
     return g.as_graph_def()
